[Builder] add PropertyBuilder
